New project structure.
Structuring the project for easier management. Preparing it for expansion.
Adding appropriate comments and python -m pip install --upgrade pip in yml.
