+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__/
 venv/
 notes.txt
 test_data/
+useful_files/
